+++ src/engine/sparqlExpressions/RelationalExpressions.h
@@ -114,7 +114,7 @@ using InExpression = relational::InExpression;
 // This function is a helper for the query planner. It allows unpacking a FILTER
 // of geof:distance and a constant. This needs to be declared in this module,
 // because the definitions of the relational expressions are hidden in its .cpp.
-std::optional<std::pair<sparqlExpression::GeoFunctionCall, size_t>>
+std::optional<std::pair<sparqlExpression::GeoFunctionCall, double>>
 getGeoDistanceFilter(const SparqlExpression& expr);
 
 }  // namespace sparqlExpression
